+++ DTO/ReceituarioDTO.cs
@@ -6,7 +6,7 @@ namespace authentication_jwt.DTO
         public int? Frequencia { get; set; }
         public string? Tempo { get; set; }
         public string? Periodo { get; set; }
-        public int? Dose { get; set; }
+        public decimal? Dose { get; set; }
         public long? UsuarioId { get; set; }
         public MedicamentoDTO? Medicamento { get; set; }
     }
